# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {
            "name": "jy.son toy project API",
            "email": "jy.son@konai.com"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
...
        "/book/:id": {
            "get": {
                "description": "Find a book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Find book",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}